+++ mot/deep_sort/detection.py
@@ -29,7 +29,7 @@ class Detection(object):
     """
 
     def __init__(self, tlwh, confidence, class_name, feature):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        self.tlwh = np.asarray(tlwh, dtype=np.float32)
         self.confidence = float(confidence)
         self.class_name = class_name
         self.feature = np.asarray(feature, dtype=np.float32)
